refactor: DB 접근 객체와 게임 객체로 분리
