+++ app/src/main/java/woowacourse/omok/database/DbHelper.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.database
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
@@ -23,22 +23,4 @@ class DbHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME, null
     override fun onDowngrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
         onUpgrade(db, oldVersion, newVersion)
     }
-}
-
-object BoardContract {
-    const val TABLE_NAME_BOARD = "board"
-
-    const val COLUMN_NAME_COLOR = "color"
-    const val COLUMN_NAME_POSITION_ROW = "row"
-    const val COLUMN_NAME_POSITION_COLUMN = "column"
-    const val COLUMN_NAME_ID = BaseColumns._ID
-
-    const val SQL_CREATE_BOARD_ENTERIES =
-        "CREATE TABLE $TABLE_NAME_BOARD (" +
-                "${COLUMN_NAME_ID} INTEGER PRIMARY KEY AUTOINCREMENT," +
-                "$COLUMN_NAME_COLOR TEXT," +
-                "$COLUMN_NAME_POSITION_ROW INTEGER, " +
-                "$COLUMN_NAME_POSITION_COLUMN INTEGER)"
-
-    const val SQL_DELETE_BOARD_ENTRIES = "DROP TABLE IF EXISTS $TABLE_NAME_BOARD"
 }
